refactor(cli): extract MCP initialization into mcp-setup.ts
Implements Round 7 Candidate #4.

Creates src/cli/mcp-setup.ts with setupMcpServers() function that handles
creation of McpManager, adding servers, registering MCP tools, and
setting the connected-server count on the status line.

shared.ts setupTools() now calls setupMcpServers() instead of inlining
all the MCP startup logic.
